requirements.txt
-i https://pypi.org/simple
azure-common==1.1.28
azure-core==1.30.0 ; python_version >= '3.7'
azure-functions==1.18.0
azure-identity==1.15.0
azure-mgmt-core==1.4.0 ; python_version >= '3.7'
azure-mgmt-resource==23.0.1
certifi==2024.2.2 ; python_version >= '3.6'
cffi==1.16.0 ; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
cryptography==42.0.2 ; python_version >= '3.7'
idna==3.6 ; python_version >= '3.5'
isodate==0.6.1
msal==1.26.0 ; python_version >= '2.7'
msal-extensions==1.1.0 ; python_version >= '3.7'
packaging==23.2 ; python_version >= '3.7'
portalocker==2.8.2 ; platform_system != 'Windows'
pycparser==2.21
pyjwt[crypto]==2.8.0 ; python_version >= '3.7'
requests==2.31.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==4.9.0 ; python_version >= '3.8'
urllib3==2.2.0 ; python_version >= '3.8'
